+++ sdk/appconfiguration/app-configuration/src/models.ts
@@ -6,10 +6,10 @@ import { FeatureFlagValue } from "./featureFlag";
 import { CommonClientOptions, OperationOptions } from "@azure/core-client";
 import { SecretReferenceValue } from "./secretReference";
 import {
-  CompositionType,
+  SnapshotComposition,
   ConfigurationSettingsFilter,
-  Snapshot,
-  SnapshotStatus,
+  ConfigurationSnapshot,
+  ConfigurationSnapshotStatus,
 } from "./generated/src";
 
 /**
@@ -63,11 +63,22 @@ export const AppConfigurationApiVersion = {
 /**
  * Fields that uniquely identify a configuration setting
  */
-export interface ConfigurationSettingId extends ConfigurationSettingsFilter {
+export interface ConfigurationSettingId {
   /**
    * The etag for this setting
    */
   etag?: string;
+  /**
+   * The key for this setting.
+   * Feature flags must be prefixed with `.appconfig.featureflag/<feature-flag-name>`.
+   */
+  key: string;
+
+  /**
+   * The label for this setting. Leaving this undefined means this
+   * setting does not have a label.
+   */
+  label?: string;
 }
 /**
  * Necessary fields for updating or creating a new configuration setting
@@ -205,7 +216,7 @@ export interface OptionalSnapshotFields {
   /**
    * Which fields to return for each ConfigurationSetting
    */
-  fields?: (keyof Snapshot)[];
+  fields?: (keyof ConfigurationSnapshot)[];
 }
 
 /**
@@ -343,12 +354,7 @@ export interface ListSettingsOptions extends OptionalFields {
  */
 export interface ListConfigurationSettingsForSnapshotOptions
   extends OperationOptions,
-    OptionalFields {
-  /**
-   * Requests the server to respond with the state of the resource at the specified time.
-   */
-  acceptDateTime?: Date;
-}
+    OptionalFields {}
 
 /**
  * Options for listConfigurationSettings that allow for filtering based on keys, labels and other fields.
@@ -366,7 +372,7 @@ export interface ListSnapshots extends OptionalSnapshotFields {
   nameFilter?: string;
 
   /** Used to filter returned snapshots by their status property. */
-  statusFilter?: SnapshotStatus[];
+  statusFilter?: ConfigurationSnapshotStatus[];
 }
 
 /**
@@ -406,11 +412,11 @@ export interface ListConfigurationSettingPage
 /**
  * A page of configuration settings and the corresponding HTTP response
  */
-export interface ListSnapshotsPage extends HttpResponseField<SyncTokenHeaderField>, PageSettings {
+export interface ListSnapshotsPage extends SyncTokenHeaderField, PageSettings {
   /**
    * The configuration settings for this page of results.
    */
-  items: Snapshot[];
+  items: ConfigurationSnapshot[];
 }
 
 /**
@@ -471,15 +477,12 @@ export interface CreateSnapshotOptions extends OperationOptions {
 /**
  * Response from adding a Snapshot.
  */
-export interface SnapshotResponse extends Snapshot, SyncTokenHeaderField {}
+export interface SnapshotResponse extends ConfigurationSnapshot, SyncTokenHeaderField {}
 
 /**
  * Options used when getting a Snapshot.
  */
-export interface GetSnapshotOptions
-  extends OperationOptions,
-    HttpOnlyIfChangedField,
-    OptionalSnapshotFields {}
+export interface GetSnapshotOptions extends OperationOptions, OptionalSnapshotFields {}
 
 /**
  * Response from getting a Snapshot.
@@ -487,9 +490,14 @@ export interface GetSnapshotOptions
 export interface GetSnapshotResponse extends SnapshotResponse {}
 
 /**
- * Options used when upadting a Snapshot.
+ * Options used when updating a Snapshot.
  */
-export interface UpdateSnapshotOptions extends HttpOnlyIfUnchangedField, OperationOptions {}
+export interface UpdateSnapshotOptions extends OperationOptions {
+  /**
+   * The etag for this snapshot
+   */
+  etag?: string;
+}
 
 /**
  * Response from updating a Snapshot.
@@ -509,33 +517,18 @@ export interface SnapshotInfo {
   /** A list of filters used to filter the key-values included in the snapshot. */
   filters: ConfigurationSettingsFilter[];
   /** The composition type describes how the key-values within the snapshot are composed. The 'all' composition type includes all key-values. The 'group_by_key' composition type ensures there are no two key-values containing the same key. */
-  compositionType?: CompositionType;
+  compositionType?: SnapshotComposition;
   /** The amount of time, in seconds, that a snapshot will remain in the archived state before expiring. This property is only writable during the creation of a snapshot. If not specified, the default lifetime of key-value revisions will be used. */
-  retentionPeriod?: number;
+  retentionPeriodInSeconds?: number;
   /** The tags of the snapshot. */
   tags?: { [propertyName: string]: string };
 }
 
-/**
- * Fields for the snapshot
- */
-export interface SnapshotId {
-  /**
-   * The name for this snapshot
-   */
-  name: string;
-  /**
-   * The etag for this snapshot
-   */
-  etag?: string;
-}
-
 export {
-  State,
-  Snapshot,
+  ConfigurationSnapshot,
   ConfigurationSettingsFilter,
-  CompositionType,
-  KnownCompositionType,
-  KnownSnapshotStatus,
-  SnapshotStatus,
+  SnapshotComposition,
+  KnownSnapshotComposition,
+  KnownConfigurationSnapshotStatus,
+  ConfigurationSnapshotStatus,
 } from "./generated/src";